+++ app/src/main/AndroidManifest.xml
@@ -8,19 +8,19 @@
         android:label="@string/app_name"
         android:supportsRtl="true"
         android:theme="@style/Theme.AppCompat.Light.NoActionBar">
-        <activity android:name=".LoginActivity">
+        <activity android:name=".authentication.LoginActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".RegistrationActivity" />
-        <activity android:name=".SeekerMainActivity" />
-        <activity android:name=".OwnerMainActivity" />
-        <activity android:name=".SetDefaultModeActivity"></activity>
+        <activity android:name=".authentication.RegistrationActivity" />
+        <activity android:name=".seeker.SeekerMainActivity" />
+        <activity android:name=".owner.OwnerMainActivity" />
+        <activity android:name=".authentication.SetDefaultModeActivity"></activity>
 
         <activity
-            android:name=".PaymentActivity"
+            android:name=".mutual.PaymentActivity"
             android:label="Payment"
             android:theme="@style/AppBarActivity">
         </activity>
