+++ app/src/main/java/edu/usc/parknpay/Database/ParkingSpot.java
@@ -1,6 +1,6 @@
 // To be used by Activities when building up a query to the DatabaseTalker
 
-package edu.usc.parknpay;
+package edu.usc.parknpay.database;
 
 import java.util.Date;
