+++ app/src/main/java/edu/usc/parknpay/Owner/OwnerMainActivity.java
@@ -1,13 +1,15 @@
-package edu.usc.parknpay;
+package edu.usc.parknpay.owner;
 
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
+
+import edu.usc.parknpay.R;
 
 public class OwnerMainActivity extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_owner_main);
+        setContentView(R.layout.owner_main);
     }
 }
