+++ app/src/main/java/edu/usc/parknpay/Seeker/SeekerMainActivity.java
@@ -1,13 +1,15 @@
-package edu.usc.parknpay;
+package edu.usc.parknpay.seeker;
 
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
+
+import edu.usc.parknpay.R;
 
 public class SeekerMainActivity extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_seeker_main);
+        setContentView(R.layout.seeker_main);
     }
 }
