+++ app/src/main/res/layout/payment_info.xml
