+++ app/src/main/res/layout/registration.xml
