+++ app/src/main/res/layout/seeker_main.xml
@@ -8,6 +8,6 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="edu.usc.parknpay.SeekerMainActivity">
+    tools:context="edu.usc.parknpay.seeker.SeekerMainActivity">
 
 </RelativeLayout>
